Separate storybook for provider sections
