--- src/stories/GifPicker.stories.tsx
@@ -1,72 +1,0 @@
-import { Meta } from '@storybook/react-webpack5';
-import React from 'react';
-import { expect, userEvent, waitFor, within } from 'storybook/test';
-import { GifPicker, Theme } from '..';
-import { Tenor } from '../providers/tenor';
-
-export default {
-	title: 'Library/GifPicker',
-	component: (props: any) => (
-		<GifPicker
-			{...props}
-			provider={Tenor(props.apiKey)}
-		/>
-	),
-	argTypes: {
-		apiKey: {
-			type: { name: 'string' },
-		},
-		onGifClick: {
-			action: 'GIF selected',
-		},
-		theme: {
-			options: ['dark', 'light', 'auto'],
-			control: {
-				type: 'radio',
-			},
-		},
-		width: {
-			type: { name: 'string' },
-		},
-		height: {
-			type: { name: 'string' },
-		},
-		categoryHeight: {
-			type: { name: 'string' },
-		},
-	},
-} as Meta<typeof GifPicker>;
-
-export const Home = {
-	args: {
-		apiKey: process.env.STORYBOOK_TENOR_TOKEN,
-	},
-};
-
-export const DarkTheme = {
-	...Home,
-	args: {
-		...Home.args,
-		theme: Theme.DARK,
-	},
-};
-
-export const HomeCategory = {
-	...Home,
-	play: async ({ canvasElement }: any) => {
-		const canvas = within(canvasElement);
-
-		await waitFor(() => expect(canvas.getAllByTestId('gpr-category')[0]).toBeInTheDocument());
-		await userEvent.click(canvas.getAllByTestId('gpr-category')[1]);
-	},
-};
-
-export const Trending = {
-	...Home,
-	play: async ({ canvasElement }: any) => {
-		const canvas = within(canvasElement);
-
-		await waitFor(() => expect(canvas.getAllByTestId('gpr-category')[0]).toBeInTheDocument());
-		await userEvent.click(canvas.getAllByTestId('gpr-category')[0]);
-	},
-};
